Revert "Benchmarking: Orchestration (#26)"
This reverts commit ae3f3ea11d0595987367d52eee1bdfd61ac91fa0.
