+++ .gitignore
@@ -14,14 +14,6 @@ server
 *.rdb
 
 .vscode/
-.venv/
 
 */__pycache__/
 **/__pycache__/
-
-.pytest_cache/
-*/pytest_cache/
-**/pytest_cache/
-
-*.egg-info/
-out/
