--- data/experiments/test.yaml
@@ -1,27 +1,0 @@
-name: Byte Size Experiment
-
-output_directory: byte_size_experiment/
-
-metadata:
-  description: Modelling the effect of Byte Size of values on ORTOA-tee
-
-client_config:
-  data:
-    generator: RandomIntegerGenerator
-    minimum: 10
-    maximum: 100
-    seed_size: 100000
-    num_operations: 1000
-
-  flags:
-    - name: nthreads
-      value:
-        type: int
-        minimum: 2
-        maximum: 8
-        step: 2
-
-host_config:
-  flags:
-    - name: nthreads
-      value: 1
