--- data/wip/byte_size_experiment.yaml
@@ -1,27 +1,0 @@
-name: Byte Size Experiment
-
-output_directory: out/benchmark/byte_size_experiment
-
-metadata:
-  description: Modelling the effect of Byte Size of values on ORTOA-tee
-
-client_config:
-  data:
-    generator: ByteSizeGenerator
-    n_bytes:
-      type: int
-      minimum: 100
-      maximum: 500
-      step: 20
-    seed_size: 100000
-    num_operations: 1000
-    key_access_distribution: uniform
-
-  flags:
-  - name: nthreads
-    value: 2
-
-host_config:
-  flags:
-  - name: nthreads
-    value: 1
