--- extras/README.md
@@ -1,20 +1,0 @@
-# ORTOA SDK
-
-## Installation
-
-Our package requires `python>=3.8.*` and can be initialized via `pip`:
-
-```bash
-ORTOA/ $ python3 -m venv .venv
-ORTOA/ $ . .venv/bin/activate
-ORTOA/ $ pip install -r TODO:REQUIREMENTS
-ORTOA/ $ pip install -e ./extras
-```
-
-## Tests
-
-Our unit tests are located in `extras/test/`. They can be executed via `pytest`:
-
-```bash
-ORTOA/ $ pytest extras/test/
-```
