--- extras/ortoa/benchmark/sample.yaml
@@ -1,49 +1,0 @@
-name: Sample Experiment
-
-output_directory: out/benchmark/
-
-metadata:
-  description: Sample experiment configuration :)
-
-client_config:
-  data:
-    generator:
-      type: IntegerGenerator
-    generator:
-      type: ByteSizeGenerator
-      bytes: 160
-    generator:
-      type: ByteSizeGenerator
-      bytes:
-        minimum: 100
-        maximum: 500
-        step: 20
-        
-    seed_size: 100000
-    n_operations: 1000
-    operation_key_access_distribution: uniform
-
-  data:
-    seed: data/sample_generated_seed/seed_data.csv
-    operations: data/sample_generated_seed/sample_operations.csv
-
-  flags:
-  - name: initdb
-    value: false
-  - name: pget
-    value:
-      # Start pget at 0, go until maximum 1 (inclusive), at each iteration increment by 0.1
-      minimum: 0.0
-      maximum: 1.0
-      increment: 0.1
-  - name: nthreads
-    value:
-      # Start nthreads at 1, go until maximum 16 (inclusive), at each iteration multiply by 2
-      minimum: 1
-      maximum: 16
-      multiplier: 2
-
-host_config:
-  flags:
-  - name: nthreads
-    value: 2
