--- extras/pyproject.toml
@@ -1,10 +1,0 @@
-[build-system]
-requires = [
-  "setuptools >= 40.9.0",
-]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "ortoa-sdk"
-version = "0.0.1"
-readme = "README.md"
