--- extras/test/benchmark/infrastructure/test_create_jobs.py
@@ -1,12 +1,0 @@
-# def test_basic_experiment():
-#     pass
-
-
-# def test_single_parameter_expansion():
-#     # - Ensure that I test one where a parameter gets expanded into a range
-#     pass
-
-
-# def test_multiple_parameter_expansions():
-#     # - Ensure that I test one where many parameters get expanded into many ranges
-#     pass
