--- extras/test/benchmark/infrastructure/test_experiment_collection.py
@@ -1,10 +1,0 @@
-# def test_file_paths():
-#     pass
-
-
-# def test_directories():
-#     pass
-
-
-# def test_combination():
-#     pass
