--- extras/test/benchmark/infrastructure/test_experiment_datagen.py
@@ -1,9 +1,0 @@
-# # Again, not sure how I'll test this but I will
-
-
-# def test_basic_data_generation():
-#     pass
-
-
-# def test_range_parameter_data_generation():
-#     pass
