--- extras/test/benchmark/infrastructure/test_load_experiments.py
@@ -1,1 +1,0 @@
-# Not sure how I'll test this yet but I will
