--- extras/test/benchmark/interface/test_files/byte_size_datagen_int_multiple.yaml
@@ -1,9 +1,0 @@
-generator: ByteSizeGenerator
-n_bytes:
-  type: int
-  minimum: 100
-  maximum: 500
-  multiplier: 2
-seed_size: 100
-num_operations: 20
-key_access_distribution: uniform
