--- extras/test/benchmark/interface/test_files/random_integer_datagen.yaml
@@ -1,6 +1,0 @@
-generator: RandomIntegerGenerator
-minimum: 10
-maximum: 100
-seed_size: 100
-num_operations: 20
-key_access_distribution: uniform
