--- extras/test/benchmark/interface/test_parameters.py
@@ -1,6 +1,0 @@
-# def test_individual_parameter_deserialization():
-#     pass
-
-
-# def test_parameter_discrimination():
-#     pass
