--- scripts/test/run_benchmark_tests.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-
-REPO_ROOT=$(git rev-parse --show-toplevel)
-BENCHMARK_TEST_DIR="${REPO_ROOT}/extras/test/benchmark/"
-
-run_unit_tests() {
-    python3 -m pytest \
-        "${BENCHMARK_TEST_DIR}"
-}
